fix: cts & cpi
