+++ inc/child-page-index/block/block.json
@@ -1,0 +1,17 @@
+{
+	"apiVersion": 2,
+	"name": "vk-blocks/child-page-index",
+	"title": "Child Page Index",
+	"category": "veu-block",
+	"description": "Page List of Child Paged",
+	"textdomain": "vk-blocks",
+	"attributes": {
+		"postId": {
+			"type": "number",
+			"default": -1
+		}
+	},
+	"supports": {
+		"className": true
+	}
+}
