+++ package.json
@@ -8,6 +8,7 @@
 		"@playwright/test": "^1.30.0",
 		"@wordpress/babel-plugin-makepot": "^4.4.0",
 		"@wordpress/env": "^5.10.0",
+		"@wordpress/scripts": "^25.3.0",
 		"babel-plugin-transform-react-jsx": "^6.24.1",
 		"child_process": "^1.0.2",
 		"gulp": "^4.0.2",
@@ -34,7 +35,8 @@
 		"dist:org": "composer install --optimize-autoloader --prefer-dist --no-dev",
 		"gulp": "gulp",
 		"watch": "gulp watch",
-		"build": "npx gulp text-domain && npx gulp build",
+		"build": "npx gulp text-domain && npm run build:block && npx gulp build",
+		"build:block": "wp-scripts build --config webpack.config.js",
 		"phpunit": "npx wp-env run phpunit 'phpunit -c /var/www/html/wp-content/plugins/vk-all-in-one-expansion-unit/.phpunit.xml --verbose'"
 	},
 	"repository": {
