+++ vkExUnit.php
@@ -4,7 +4,8 @@
  * Plugin URI: https://ex-unit.nagoya
  * Description: This plug-in is an integrated plug-in with a variety of features that make it powerful your web site. Many features can be stopped individually. Example Facebook Page Plugin,Social Bookmarks,Print OG Tags,Print Twitter Card Tags,Print Google Analytics tag,New post widget,Insert Related Posts and more!
  * Version: 9.86.2.0
- * Requires PHP: 7.2
+ * Requires PHP: 7.4
+ * Requires at least: 5.9
  * Author: Vektor,Inc.
  * Text Domain: vk-all-in-one-expansion-unit
  * Domain Path: /languages
